Set the download URL to the master branch on channel_manager.py
Currently it is partially implemented the tag management, but
this systems seems not easily feasible because some packages are
tagged based on tag prefixes. Then it is needed to also implement
on the `.gitmodules` file, the `tag_prefix` field to select which
is the tag prefix, and for which platform it it.